+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 - anonymous without chat
-- zh-TW, pt-PT
+- changed jwt implementation in csharp, csharp-mvc, php, ruby
+- added hy, eu, zh-TW, ms, pt-PT
 
 ## 1.2.0
 - ruby v3.0